Module wasmer_journal::concrete::compacting

Structs

CompactResult

Represents the results of a compaction operation

CompactingJournal

CompactingJournalRx

CompactingJournalTx

Deduplicates memory and stacks to reduce the number of volume of log events sent to its inner capturer. Compacting the events occurs in line as the events are generated

MemoryRange 🔒

State 🔒

SubGroupIndex 🔒

Index of a group of subevents in the journal which relate to a particular collective impact. For example. Creating a new file which may consist of an event to open a file, the events for writing the file data and the closing of the file are all related to a group of sub events that make up the act of creating that file. During compaction these events will be grouped together so they can be retained or discarded based on the final deterministic outcome of the entire log.

SubGroupOfevents 🔒

Subgroup of events that may or may not be retained in the final journal as it is compacted.

Type Aliases

Fd