wasmer_journal::concrete

Module filter

Structs

FilteredJournal

Filters out a specific set of journal events and drops the rest, this journal can be useful for restoring to a previous call point but retaining the memory changes (e.g. WCGI runner).

FilteredJournalBuilder

Constructs a filter with a set of parameters that will be filtered on

FilteredJournalConfig 🔒

Represents what will be filtered by the filtering process

FilteredJournalRx

FilteredJournalTx