wasmer_journal::concrete

Module log_file

Structs

LogFileJournal

The LogFile snapshot capturer will write its snapshots to a linear journal and read them when restoring. It uses the bincode serializer which means that forwards and backwards compatibility must be dealt with carefully.

LogFileJournalRx

LogFileJournalTx

TxState 🔒