wasmer_journal::concrete::transaction

Struct TransactionJournal

pub struct TransactionJournal<W: WritableJournal, R: ReadableJournal> {
    pub(super) tx: TransactionJournalTx<W>,
    pub(super) rx: TransactionJournalRx<R>,
}

Journal which will store the events locally in memory until it is either committed or rolled back

Fields
tx: TransactionJournalTx<W>rx: TransactionJournalRx<R>
Implementations

impl TransactionJournal<Box<DynWritableJournal>, Box<DynReadableJournal>>

pub fn new<J>(inner: J) -> Self
where J: Journal,

Creates a transactional journal which will hold events in memory until the journal is either committed or rolled back

impl<W: WritableJournal, R: ReadableJournal> TransactionJournal<W, R>

pub fn into_inner(self) -> RecombinedJournal<W, R>