wasmer_journal

Enum JournalEntryRecordType

#[repr(u16)]
pub enum JournalEntryRecordType {

    InitModuleV1 = 1,
    ProcessExitV1 = 2,
    SetThreadV1 = 3,
    CloseThreadV1 = 4,
    FileDescriptorSeekV1 = 5,
    FileDescriptorWriteV1 = 6,
    UpdateMemoryRegionV1 = 7,
    SetClockTimeV1 = 9,
    OpenFileDescriptorV1 = 10,
    CloseFileDescriptorV1 = 11,
    RenumberFileDescriptorV1 = 12,
    DuplicateFileDescriptorV1 = 13,
    CreateDirectoryV1 = 14,
    RemoveDirectoryV1 = 15,
    PathSetTimesV1 = 16,
    FileDescriptorSetTimesV1 = 17,
    FileDescriptorSetSizeV1 = 18,
    FileDescriptorSetFlagsV1 = 19,
    FileDescriptorSetRightsV1 = 20,
    FileDescriptorAdviseV1 = 21,
    FileDescriptorAllocateV1 = 22,
    CreateHardLinkV1 = 23,
    CreateSymbolicLinkV1 = 24,
    UnlinkFileV1 = 25,
    PathRenameV1 = 26,
    ChangeDirectoryV1 = 27,
    EpollCreateV1 = 28,
    EpollCtlV1 = 29,
    TtySetV1 = 30,
    CreatePipeV1 = 31,
    CreateEventV1 = 32,
    PortAddAddrV1 = 33,
    PortDelAddrV1 = 34,
    PortAddrClearV1 = 35,
    PortBridgeV1 = 36,
    PortUnbridgeV1 = 37,
    PortDhcpAcquireV1 = 38,
    PortGatewaySetV1 = 39,
    PortRouteAddV1 = 40,
    PortRouteClearV1 = 41,
    PortRouteDelV1 = 42,
    SocketOpenV1 = 43,
    SocketListenV1 = 44,
    SocketBindV1 = 45,
    SocketConnectedV1 = 46,
    SocketAcceptedV1 = 47,
    SocketJoinIpv4MulticastV1 = 48,
    SocketJoinIpv6MulticastV1 = 49,
    SocketLeaveIpv4MulticastV1 = 50,
    SocketLeaveIpv6MulticastV1 = 51,
    SocketSendFileV1 = 52,
    SocketSendToV1 = 53,
    SocketSendV1 = 54,
    SocketSetOptFlagV1 = 55,
    SocketSetOptSizeV1 = 56,
    SocketSetOptTimeV1 = 57,
    SocketShutdownV1 = 58,
    SnapshotV1 = 59,
    ClearEtherealV1 = 60,
    OpenFileDescriptorV2 = 61,
    DuplicateFileDescriptorV2 = 62,
    FileDescriptorSetFdFlagsV1 = 63,
    SocketPairV1 = 64,

}

Variants

InitModuleV1 = 1

ProcessExitV1 = 2

SetThreadV1 = 3

CloseThreadV1 = 4

FileDescriptorSeekV1 = 5

FileDescriptorWriteV1 = 6

UpdateMemoryRegionV1 = 7

SetClockTimeV1 = 9

OpenFileDescriptorV1 = 10

CloseFileDescriptorV1 = 11

RenumberFileDescriptorV1 = 12

DuplicateFileDescriptorV1 = 13

CreateDirectoryV1 = 14

RemoveDirectoryV1 = 15

PathSetTimesV1 = 16

FileDescriptorSetTimesV1 = 17

FileDescriptorSetSizeV1 = 18

FileDescriptorSetFlagsV1 = 19

FileDescriptorSetRightsV1 = 20

FileDescriptorAdviseV1 = 21

FileDescriptorAllocateV1 = 22

CreateHardLinkV1 = 23

CreateSymbolicLinkV1 = 24

UnlinkFileV1 = 25

PathRenameV1 = 26

ChangeDirectoryV1 = 27

EpollCreateV1 = 28

EpollCtlV1 = 29

TtySetV1 = 30

CreatePipeV1 = 31

CreateEventV1 = 32

PortAddAddrV1 = 33

PortDelAddrV1 = 34

PortAddrClearV1 = 35

PortBridgeV1 = 36

PortUnbridgeV1 = 37

PortDhcpAcquireV1 = 38

PortGatewaySetV1 = 39

PortRouteAddV1 = 40

PortRouteClearV1 = 41

PortRouteDelV1 = 42

SocketOpenV1 = 43

SocketListenV1 = 44

SocketBindV1 = 45

SocketConnectedV1 = 46

SocketAcceptedV1 = 47

SocketJoinIpv4MulticastV1 = 48

SocketJoinIpv6MulticastV1 = 49

SocketLeaveIpv4MulticastV1 = 50

SocketLeaveIpv6MulticastV1 = 51

SocketSendFileV1 = 52

SocketSendToV1 = 53

SocketSendV1 = 54

SocketSetOptFlagV1 = 55

SocketSetOptSizeV1 = 56

SocketSetOptTimeV1 = 57

SocketShutdownV1 = 58

SnapshotV1 = 59

ClearEtherealV1 = 60

OpenFileDescriptorV2 = 61

DuplicateFileDescriptorV2 = 62

FileDescriptorSetFdFlagsV1 = 63

SocketPairV1 = 64

Implementations

impl JournalEntryRecordType

pub unsafe fn deserialize_archive(self, data: &[u8]) -> Result<JournalEntry<'_>>

Safety

rykv makes direct memory references to achieve high performance however this does mean care must be taken that the data itself can not be manipulated or corrupted.