wasmer_journal

Trait ReadableJournal

pub trait ReadableJournal: Debug {
    // Required methods
    fn read(&self) -> Result<Option<LogReadResult<'_>>>;
    fn as_restarted(&self) -> Result<Box<DynReadableJournal>>;
}

The snapshot capturer will take a series of objects that represents the state of a WASM process at a point in time and saves it so that it can be restored. It also allows for the restoration of that state at a later moment

Required Methods

fn read(&self) -> Result<Option<LogReadResult<'_>>>

Returns a stream of snapshot objects that the runtime will use to restore the state of a WASM process to a previous moment in time

fn as_restarted(&self) -> Result<Box<DynReadableJournal>>

Resets the journal so that reads will start from the beginning again