wasmer_journal
Trait WritableJournal

pub trait WritableJournal: Debug {
    // Required methods
    fn write<'a>(&'a self, entry: JournalEntry<'a>) -> Result<LogWriteResult>;
    fn flush(&self) -> Result<()>;

    // Provided methods
    fn commit(&self) -> Result<usize> { ... }
    fn rollback(&self) -> Result<usize> { ... }
}

The snapshot capturer will take a series of objects that represents the state of a WASM process at a point in time and saves it so that it can be restored. It also allows for the restoration of that state at a later moment

Required Methods

fn write<'a>(&'a self, entry: JournalEntry<'a>) -> Result<LogWriteResult>

Takes in a stream of snapshot log entries and saves them so that they may be restored at a later moment

fn flush(&self) -> Result<()>

Flushes the data to disk or network

Provided Methods

fn commit(&self) -> Result<usize>

Commits the transaction Returns the number of events committed

fn rollback(&self) -> Result<usize>

Rolls back the transaction and aborts its changes Returns the number of events rolled back