wasmer_journal

Module entry

Enums

JournalEntry

Represents a log entry in a snapshot log stream that represents the total state of a WASM process at a point in time.

SocketJournalEvent

SocketOptTimeType

SocketShutdownHow

Type Aliases

Fd 🔒