wasmer_journal

Trait Journal

pub trait Journal:
    WritableJournal
    + ReadableJournal
    + AsDynReadableJournal
    + Debug {
    // Required method
    fn split(self) -> (Box<DynWritableJournal>, Box<DynReadableJournal>);
}

The snapshot capturer will take a series of objects that represents the state of a WASM process at a point in time and saves it so that it can be restored. It also allows for the restoration of that state at a later moment

Required Methods

fn split(self) -> (Box<DynWritableJournal>, Box<DynReadableJournal>)

Splits the journal into a read and write side