Trait wasmer_journal::WritableJournal

pub trait WritableJournal {
    // Required method
    fn write<'a>(&'a self, entry: JournalEntry<'a>) -> Result<LogWriteResult>;
}

The snapshot capturer will take a series of objects that represents the state of a WASM process at a point in time and saves it so that it can be restored. It also allows for the restoration of that state at a later moment

Required Methods

fn write<'a>(&'a self, entry: JournalEntry<'a>) -> Result<LogWriteResult>

Takes in a stream of snapshot log entries and saves them so that they may be restored at a later moment